Update Entity theo DB mới và xóa những thứ dư thừa

+++ src/main/java/com/example/Group5/dao/AppRoleDAO.java
@@ -16,7 +16,7 @@ public class AppRoleDAO {
     @Autowired
     private EntityManager entityManager;
 
-    public List<String> getRoleNames(Long userId) {
+    public List<String> getRoleNames(int userId) {
         String sql = "Select ur.appRole.roleName from " + UserRole.class.getName() + " ur " //
                 + " where ur.appUser.userId = :userId ";
